requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
langchain>=0.1.0
langchain-community>=0.0.10
qdrant-client>=1.7.1
sentence-transformers>=2.2.2
pandas>=2.1.3
openpyxl>=3.1.2
python-dotenv>=1.0.0
pydantic>=2.5.0
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
ollama>=0.1.7

